bug fix to tamm-tensor
Former-commit-id: 1877b048ce3a634851e927f0ff0bdbb38831031a
Former-commit-id: 9c8b8fcab0af1986fb4b03947fd162da5ef3010e
Former-commit-id: 7c2daefb557821705cfa4747ee52e4be84e3f271
Former-commit-id: 0a6657942d790b83fe8ea6a354eadc4d1e575ded [formerly c758ed044f44d15f75de5f2046049546b5846f3a]
Former-commit-id: f860c2e712a4e8ea31a53361ec9132aa1aa4ebb7

+++ tensor/gmem.h
@@ -38,7 +38,7 @@ enum Types { Int, Double };
 
 enum Operation { Plus = 0, Multiply, Max, Min, AbsMax, AbsMin };
 
-Handle create(Types type, int size, char* name);
+Handle create(Types type, int64_t size, char* name);
 uint64_t ranks();
 int64_t atomic_fetch_add(Handle handle, int pos, int amount);
 void zero(Handle handle);
